nouveau: big rewrite of nv50 kms code
This is a essentially a big squash merge from the kernel tree I've been
working in, sorry about the loss of history, but commiting each patch
individually would've taken quite a while to do.

The nv50 kms code now uses the helpers in the drm, simplifying the
code greatly.

The code is far more stable than previously on all the G8x cards I've
been able to test on, but there's likely still bugs to be ironed out!